Refactor command classes to new package structure 'xyz.gamecrash.consolepipe.Commands.ConsolePipe'
