+++ src/main/java/xyz/gamecrash/consolepipe/Commands/ConsolePipe/FilterCommand.java
@@ -1,4 +1,4 @@
-package xyz.gamecrash.consolepipe.Commands;
+package xyz.gamecrash.consolepipe.Commands.ConsolePipe;
 
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.context.CommandContext;
